yt:git_svn_commit:ffe134d63249d29c39f51e36777f7f65f7a8c7c1
yt:svn_revision:2490220

+++ yt/core/actions/future-inl.h
@@ -1101,10 +1101,10 @@ public:
         if (ShouldCompletePrematurely()) {
             SetPromisePrematurely(Promise_);
         } else {
-            for (int index = 0; index < Futures_.size(); ++index) {
+            for (size_t index = 0; index < Futures_.size(); ++index) {
                 Futures_[index].Subscribe(BIND(&TFutureCombinerBase::OnFutureSet, MakeStrong(this), index));
             }
-            Promise_.OnCanceled(BIND(&TFutureCombinerBase::CancelFutures, MakeWeak(this)));            
+            Promise_.OnCanceled(BIND(&TFutureCombinerBase::CancelFutures, MakeWeak(this)));
         }
         return Promise_;
     }
@@ -1116,7 +1116,7 @@ protected:
 
     void CancelFutures()
     {
-        for (int index = 0; index < Futures_.size(); ++index) {
+        for (size_t index = 0; index < Futures_.size(); ++index) {
             Futures_[index].Cancel();
         }
     }
@@ -1151,7 +1151,7 @@ public:
 private:
     TFutureCombinerResultHolder<T> ResultHolder_;
     std::atomic<int> PendingResponseCount_;
-    
+
 
     virtual void OnFutureSet(int futureIndex, const TErrorOr<T>& result) override
     {
@@ -1162,12 +1162,12 @@ private:
         }
 
         ResultHolder_.SetItem(futureIndex, result);
-        
+
         if (--PendingResponseCount_ == 0) {
             ResultHolder_.SetPromise(this->Promise_);
         }
     }
-    
+
     virtual bool ShouldCompletePrematurely() override
     {
         return this->Futures_.empty();
@@ -1192,7 +1192,7 @@ private:
     TFutureCombinerResultHolder<T> ResultHolder_;
     std::atomic<int> PendingResponseCount_;
     std::atomic<int> CurrentResponseIndex_ = {0};
-    
+
 
     virtual void OnFutureSet(int /*futureIndex*/, const TErrorOr<T>& result) override
     {
@@ -1206,12 +1206,12 @@ private:
         if (responseIndex < Quorum_) {
             ResultHolder_.SetItem(responseIndex, result);
         }
-        
+
         if (--PendingResponseCount_ == 0) {
             ResultHolder_.SetPromise(this->Promise_);
         }
     }
-    
+
     virtual bool ShouldCompletePrematurely() override
     {
         return this->Futures_.empty() || Quorum_ == 0;
@@ -1241,7 +1241,7 @@ private:
             this->Promise_.Set(std::move(Results_));
         }
     }
-    
+
     virtual bool ShouldCompletePrematurely() override
     {
         return this->Futures_.empty();
